+++ tslearn/tests/test_estimators.py
@@ -197,7 +197,8 @@ def test_all_estimators(name, Estimator):
                  Estimator().get_tags()["allow_nan"])
     if allow_nan:
         checks.ALLOW_NAN.append(name)
-    if name in ["ShapeletModel", "SerializableShapeletModel"]:
+    if name in ["GlobalAlignmentKernelKMeans", "ShapeletModel",
+                "SerializableShapeletModel"]:
         # Deprecated models
         return
     check_estimator(Estimator)