+++ CHANGELOG.md
@@ -18,6 +18,8 @@ consistent clustering runs (helps avoid empty cluster situations)
 completely sklearn-compliant
 * The `shapelets` module now requires tensorflow>=2 as dependency (was keras
 + tensorflow==1.* up to version 0.3)
+* `GlobalAlignmentKernelKMeans` is deprecated in favor of `KernelKMeans` that
+accepts various kernels (and "gak" is the default)
 * `ShapeletModel` is now called `LearningShapelets` to be more explicit about
 which shapelet-based classifier is implemented. `ShapeletModel` is still
 available as an alias, but is now considered part of the private API